Implement name tagging
