+++ packages/fileops/package.json
@@ -15,10 +15,13 @@
   "dependencies": {
     "@common/common": "^1.0.0",
     "dateformat": "^4.6.3",
-    "exif": "^0.6.0"
+    "exif": "^0.6.0",
+    "ffmpeg-static": "^5.2.0",
+    "fluent-ffmpeg": "^2.1.3"
   },
   "devDependencies": {
     "@types/dateformat": "^5.0.2",
+    "@types/fluent-ffmpeg": "^2.1.24",
     "@types/jest": "^29.5.12",
     "jest": "^29.7.0",
     "ts-jest": "^29.2.3"
